# internal/docs/swagger.yaml
basePath: /api
definitions:
  httpsuite.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  models.MessageRate:
    properties:
      rate:
        type: number
    type: object
  models.Metadata:
    properties:
      default_queue_type:
        type: string
      description:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  models.RMQQueue:
    properties:
      consumers:
        type: integer
      deliver_rate:
        type: number
      message_bytes:
        type: integer
      message_bytes_persistent:
        type: integer
      messages:
        type: integer
      messages_unacknowledged:
        type: integer
      name:
        type: string
      publish_rate:
        type: number
      redeliver_rate:
        type: number
      vhost:
        type: string
    type: object
  models.RMQVhostLimits:
    properties:
      max_connections:
        type: integer
      max_queues:
        type: integer
      vhost:
        type: string
    type: object
  models.RMQVhostUsage:
    properties:
      disk_bytes:
        type: integer
      message_bytes:
        type: integer
      name:
        type: string
    type: object
...
  models.Vhost:
    properties:
      cluster_state:
        additionalProperties:
          type: string
        type: object
      default_queue_type:
        type: string
      description:
        type: string
      messages:
        type: integer
      messages_details:
        $ref: '#/definitions/models.MessageRate'
      messages_ready:
        type: integer
      messages_ready_details:
        $ref: '#/definitions/models.MessageRate'
      messages_unacknowledged:
        type: integer
      messages_unacknowledged_details:
        $ref: '#/definitions/models.MessageRate'
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        type: string
      protected_from_deletion:
        type: boolean
      tags:
        items:
          type: string
        type: array
      tracing:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Norsk helsenett SF
    url: https://github.com/NorskHelsenett/unimq-dashboard
  description: API for RabbitMQ Dashboard application.
  title: RabbitMQ Dashboard API
  version: "1.0"
paths:
...
  /v1/vhosts:
    get:
      description: Get a list of all vhosts in the RabbitMQ cluster
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Vhost'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get all vhosts
      tags:
      - Vhosts
  /v1/vhosts/{vhost-name}:
    get:
      description: Get details of a specific vhost by name
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vhost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get vhost details
      tags:
      - Vhosts
  /v1/vhosts/{vhost-name}/limits:
    get:
      description: Get limits of a specific vhost by name
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RMQVhostLimits'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get vhost limits
      tags:
      - Vhosts
...
  /v1/vhosts/{vhost-name}/queues:
    get:
      description: Fetches a list of all queues in a specified virtual host.
      parameters:
      - description: Virtual Host
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: HTML page with queue metrics
          schema:
            items:
              $ref: '#/definitions/models.RMQQueue'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get Queues for a specific vhost
      tags:
      - Queues
  /v1/vhosts/{vhost-name}/queues/{queue-id}:
    get:
      description: Fetches details of all queues in a specified virtual host.
      parameters:
      - description: Virtual Host
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Queue Name
        in: path
        name: queue-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of queue details
          schema:
            items:
              $ref: '#/definitions/models.RMQQueue'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get a specific queue to a specified virtual host
      tags:
      - Queues
  /v1/vhosts/{vhost-name}/usage:
    get:
      description: Get usage statistics for a specific vhost in the RabbitMQ cluster
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RMQVhostUsage'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get Vhost usage
      tags:
      - Vhosts
securityDefinitions:
  OAuth2:
    authorizationUrl: http://localhost:5556/dex/auth
    flow: accessCode
    scopes:
      audience:server:client_id:unimq-dashboard: Issue a token whose audience the
        dashboard API accepts
      email: User email address
      groups: Group membership used for authorization
      openid: OpenID Connect identity
      profile: User profile
    tokenUrl: http://localhost:5556/dex/token
    type: oauth2
  bearer:
    description: '"JWT token for authentication, obtained from the Dex OIDC provider,
      formatted as ''Bearer {token}'' in the Authorization header"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"